pyproject.toml
[project]
name = "autoshift"
version = "0.1.0"
description = ""
authors = [
    {name = "Vsevolod Martynov"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
